requirements.txt
streamlit>=1.28.0
googlemaps==4.10.0
requests==2.31.0
pandas>=2.0.0,<2.2.0
numpy>=1.24.0,<1.26.0
python-dotenv==1.0.0
aiohttp==3.8.5
asyncio-throttle==1.0.2
openai>=1.0.0 